Add NamedModel
A new base model makes it easy to manage models that require a name and
related fields. Existing models have been modified to use this base
model without breaking backwards compatibility, i.e. no actual database
fields have been changed during consolidation.
