+++ netbox_inventory/models/assets.py
@@ -4,7 +4,7 @@ from django.db import models
 from django.forms import ValidationError
 from django.urls import reverse
 
-from netbox.models import NestedGroupModel, NetBoxModel
+from netbox.models import NestedGroupModel
 from netbox.models.features import ImageAttachmentsMixin
 
 from ..choices import AssetStatusChoices, HardwareKindChoices
@@ -15,9 +15,10 @@ from ..utils import (
     get_prechange_field,
     get_status_for,
 )
+from .mixins import NamedModel
 
 
-class InventoryItemGroup(NestedGroupModel):
+class InventoryItemGroup(NestedGroupModel, NamedModel):
     """
     Inventory Item Groups are groups of simmilar InventoryItemTypes.
     This allows you to, for example, have one Group for all your 10G-LR SFP
@@ -26,8 +27,6 @@ class InventoryItemGroup(NestedGroupModel):
     """
 
     slug = None  # remove field that is defined on NestedGroupModel
-
-    comments = models.TextField(blank=True)
 
     class Meta:
         ordering = ['name']
@@ -47,11 +46,13 @@ class InventoryItemGroup(NestedGroupModel):
         return reverse('plugins:netbox_inventory:inventoryitemgroup', args=[self.pk])
 
 
-class InventoryItemType(NetBoxModel, ImageAttachmentsMixin):
+class InventoryItemType(NamedModel, ImageAttachmentsMixin):
     """
     Inventory Item Type is a model (make, part number) of an Inventory Item. In
     that it is simmilar to Device Type or Module Type.
     """
+
+    name = None  # remove field that is defined on PrimaryModel
 
     manufacturer = models.ForeignKey(
         to='dcim.Manufacturer',
@@ -78,13 +79,6 @@ class InventoryItemType(NetBoxModel, ImageAttachmentsMixin):
         null=True,
         verbose_name='Inventory Item Group',
     )
-    description = models.CharField(
-        max_length=200,
-        blank=True,
-    )
-    comments = models.TextField(
-        blank=True,
-    )
 
     clone_fields = [
         'manufacturer',
@@ -104,7 +98,7 @@ class InventoryItemType(NetBoxModel, ImageAttachmentsMixin):
         return reverse('plugins:netbox_inventory:inventoryitemtype', args=[self.pk])
 
 
-class Asset(NetBoxModel, ImageAttachmentsMixin):
+class Asset(NamedModel, ImageAttachmentsMixin):
     """
     An Asset represents a piece of hardware we want to keep track of. It has a
     make (model, part number) that is one of: Device Type, Module Type,
@@ -128,10 +122,6 @@ class Asset(NetBoxModel, ImageAttachmentsMixin):
         blank=True,
         null=False,
         default='',
-    )
-    description = models.CharField(
-        max_length=200,
-        blank=True,
     )
     asset_tag = models.CharField(
         help_text='Identifier assigned by owner',
@@ -292,10 +282,6 @@ class Asset(NetBoxModel, ImageAttachmentsMixin):
         blank=True,
         null=True,
         verbose_name='Warranty End',
-    )
-
-    comments = models.TextField(
-        blank=True,
     )
 
     clone_fields = [
