+++ pyproject.toml
@@ -28,6 +28,11 @@ include-package-data = true
 [tool.setuptools.packages.find]
 include = ["netbox_inventory*"]
 
+[tool.ruff]
+exclude = [
+    "netbox_inventory/migrations",
+]
+
 [tool.ruff.lint]
 extend-select = ["E4", "E7", "E9", "F", "W", "C"]
 ignore = ["F403", "F405"]
